Make use of jsonb columns for the resource -> floored versions mapping
We were storing the resources and their floored versions in two comma
separated strings, which was quite frankly, grim. This is so much nicer.
We now store the data in a single jsonb column called
"resources_and_versions". The keys are the resource ids and the values
are their floored versions. Checking for query doi entries which are
contain data from a given resource then can be completed by running some
SQL like this:

  select
    *
  from query_dois
  where resources_and_versions->>'<resource_id>' is not null;

which is much much nicer.

To achieve this in the current version of ckan we're using, a jsonb
custom type has been added as the sqlalchemy version we're forced to use
doesn't support it yet.