Fix QueryDOI.on_resource function
This function was broken and although it didn't error, it never matched
any resources when used. Now that the JSONB column type is available
(thanks to a bump in sqlalchemy version from the ckan upgrade) we can
switch the column type to JSONB and use the has_key function create a
filter.