Don't produce html and try and insert it into a tag's attribute
I think this must have been a layover from previously using the author
names in the citation text. We should be identifying all the authors in
citation meta tags though so we don't need to use any et als and we
don't need to produce a truncated list with some funky hover stuff.