Remove lines that aren't needed with a newer version of sqlalchemy
We've internally upgraded to 2.7.x for the ubuntu 18.04 upgrade as
we've gone up from postgresql 9 -> 11. This means we need to remove
this load function as psycopg2 seems to be smart enough to know that
the data should be loaded and returned as a dict, even though it
doesn't seem to be smart enough to recognise that it should dump a
dict automatically. This won't be needed with ckan 2.8 as that is using
an even higher version of sqlalchemy which has a JSONB type.