Implement multisearch minting and add multisearch download hook
There's quite a lot in this commit, apologies :(

To start with, we now have the fairly nasty situation where we
support queries that are for the old datastore_search action and
the new datastore_multisearch action. To accommodate for this
the query_dois table has had an extra column added which stores
the query version. If the query doi is for a datastore_search
search then this is left null, otherwise it is filled in. This
choice has been made to make supporting both the new multisearch
and the old search easier as users may get confused if the doi
they minted on the old search action now resolves through the
landing page to the new search page. Once we migrate completely
over and rely entirely on the new search page we'll migrate all
of these old dois to the new query schema.

A minting action called `create_doi` has been created which allows
the creation of multisearch dois without triggering a download.

For normal multisearch downloads, a versioned datastore extension
hook is used, allowing us to create a DOI and add it to the email's
body text.

NOTE: you will need to run the database migration before you run
this code as there is a new column in the query_dois table.