feat: add the total dois to the query doi sidebar
A total number of DOIs is now shown in the recent DOIs sidebar section. Additionally, this commit adds an explanation to go along with the total. This is partly there because it's actually quite useful to explain what's going on and partly because it helped with styling/formatting where the total was made visible to the user in the sidebar.