feat: compatibility with vds vNext
BREAKING CHANGE: compatibility with vds vNext

Main change here it that resources and versions are purely an internal concept now just stored in the database not passed to the user or expected from the user. This is because with the new vds changes we don't actually need to store this information to recreate a search across multiple resources. In the past, vds allowed new resource versions to be created as long as they were greater than the previous version of a resource which forced us to store the resources and versions to avoid creating DOIs which could reference data created after the DOI is minted but with a version before it. Because this is no longer possible, we don't need to worry about this any more. Currently, the database still stores these resource and version dicts but it's strictly not necessary (though it does make matching queries easier). This might get changed in the future (we could achieve the same by storing a requested version and a single rounded version for all resources) but for now it's not a problem.