fix: count non-datastore resources as inacccessible (multisearch)
Rarely, resource data may get deleted from the datastore, but the resource itself is still accessible as a raw file. This fix covers that eventuality, and also adds a resource breakdown for missing data (where possible) on the landing page.

This does rely on datastore_active being set properly, mostly for speed reasons.

Closes: #66